further updates.  it now kind of works!

+++ js/services.js
@@ -110,13 +110,14 @@ function($window, $http, $cookies, $rootScope) {             // service construc
  */
 factory('noteService',
         function($http, $resource, $window, $rootScope, localStorageService, sessionService) {
-            var KEY = "add key here";
+            var KEY = "your api key here";
             var BASE_PATH = 'https://api.mongolab.com/api/1/databases/simple_test_db/collections/'
             var noteList;
             var session = sessionService.getSession();
 
 
-            var getNoteList = function() {
+            var getNoteList = function(getRemotes) {
+                console.log("getNoteList()  getRemotes = ", getRemotes);
                 var localNotes = localStorageService.get('localNotes');
 
                 if (!localNotes) {
@@ -124,8 +125,9 @@ factory('noteService',
                 }
 
                 noteList = localNotes;
+                console.log("local notes:", noteList);
 
-                if (session) {
+                if (session && getRemotes) {
                     var url = BASE_PATH + session.name;
                     $http.get(url,
                               {                                                       // this is a configuration object.  at the moment we only use the params field
@@ -135,7 +137,7 @@ factory('noteService',
 //                                    console.log("success --> data: ", data, "; status: ", status);
                                     for (var i = 0; i < data.length; i++) {
                                         var note = angular.fromJson(data[i].note);
-                                        note.remoteId = data[i]._id;
+                                        note.remoteId = data[i]._id.$oid;
                                         var found;
                                         for (var j = 0; j < localNotes.length; j++) {
                                             if (localNotes[j].id == note.id) {
@@ -154,12 +156,15 @@ factory('noteService',
                                         console.log("found note: ", note);
                                     }
                                     noteList = localNotes;
+                                    // save local note list with remotes too
+                                    localStorageService.add('localNotes', noteList);
+
+                                    // broadcast update
                                     $rootScope.$broadcast('remoteNotesUpdated', noteList);
                                 }).
                         error(function(data, status) {
                                   console.log("error --> data: ", data, "; status: ", status, "; error count: ", errCount);
                               });
-
                 }
 
                 return noteList;
@@ -167,7 +172,7 @@ factory('noteService',
 
 
             var getNote = function(noteId) {
-                getNoteList();
+                getNoteList(false);
                 var note = null;
                 if (noteList) {
                     for (var i = 0; i < noteList.length; i++) {
@@ -179,19 +184,19 @@ factory('noteService',
                 }
 
                 if (note == null) {
+                    var newId = new Date().getTime();
                     note = {
-                        id: new Date().getTime(),
+                        id: newId,
                         date: new Date(),
-                        content: ''
+                        content: 'New Note...'
                     };
+                    console.log("created new note: ", note);
                 }
                 return note;
             };
 
-
-            var saveNote = function(note) {
+            function saveNoteLocally(note) {
                 var found = false;
-                getNoteList();
                 if (noteList) {
                     for (var i = 0; i < noteList.length; i++) {
                         if (noteList[i].id == note.id) {
@@ -206,34 +211,62 @@ factory('noteService',
                 } else {
                     noteList = [ note ];
                 }
+                localStorageService.add('localNotes', noteList);
+                console.log("stored local notes list: ", noteList);
+            }
+
+            var saveNote = function(note) {
+                saveNoteLocally(note);
 
                 if (session) {
-                    var url = BASE_PATH + session.name + '?apiKey=' + KEY;
-                    $http.post(url,
-                              { note: angular.toJson(note) },
-                              {
-                                  params: { apiKey: KEY }
-                              }
-                    ).
-                        success(function(data, status) {           // on success method
-                                    console.log("success posting --> data: ", data, "; status: ", status);
-                                }).
-                        error(function(data, status) {
-                                  console.log("error posting --> data: ", data, "; status: ", status);
-                              });
+                    var url = BASE_PATH + session.name;
+                    if (note.remoteId) {
+                        url += '/' + note.remoteId;
+                        $http.put(url,
+                                   { note: angular.toJson(note) },
+                                   {
+                                       params: { apiKey: KEY }
+                                   }
+                        ).
+                            success(function(data, status) {           // on success method
+                                        console.log("saved note: ", note);
+                                    }).
+                            error(function(data, status) {
+                                      console.log("error saving --> data: ", data, "; status: ", status);
+                                  });
 
+                    } else {
+                        $http.post(url,
+                                   { note: angular.toJson(note) },
+                                   {
+                                       params: { apiKey: KEY }
+                                   }
+                        ).
+                            success(function(data, status) {           // on success method
+                                        note.remoteId = data._id.$oid;
+
+                                        // update with remote id
+                                        saveNoteLocally(note);
+                                        console.log("saved note: ", note);
+                                    }).
+                            error(function(data, status) {
+                                      console.log("error saving --> data: ", data, "; status: ", status);
+                                  });
+
+                    }
                 }
-
-                localStorageService.add('localNotes', noteList);
             };
 
 
             var deleteNote = function(note) {
+                console.log("deleting note: ", note);
                 var nl = [];
-                getNoteList();
+                getNoteList(false);
                 if (noteList) {
                     for (var i = 0; i < noteList.length; i++) {
-                        if (noteList[i].id != note.id) {
+                        var oldNote = noteList[i];
+                        console.log("oldNote: ", oldNote, ", new note: ", note);
+                        if (oldNote.id != note.id) {
                             nl.push(noteList[i]);
                         }
                     }
@@ -241,26 +274,26 @@ factory('noteService',
                     localStorageService.add('localNotes', nl);
                 }
 
-                if (session) {
-                    var url = BASE_PATH + session.name + '/' + note.remoteId.$oid;
+                if (session && note.remoteId) {
+                    var url = BASE_PATH + session.name + '/' + note.remoteId;
                     $http.delete(url,
                                {
                                    params: {
-                                       apiKey: KEY,
+                                       apiKey: KEY
                                    }
                                }
                     ).
                         success(function(data, status) {           // on success method
-                                    console.log("success putting --> data: ", data, "; status: ", status);
+                                    // broadcast update
+                                    $rootScope.$broadcast('noteDeleted', true);
+                                    console.log("success deleting note: ", note, " --> data: ", data, "; status: ", status);
                                 }).
                         error(function(data, status) {
-                                  console.log("error putting --> data: ", data, "; status: ", status);
+                                  console.log("error deleting --> data: ", data, "; status: ", status);
                               });
-
+                } else {
+                    $rootScope.$broadcast('noteDeleted', false);
                 }
-
-                $rootScope.$broadcast('noteDeleted', [ note ]);
-                note = null;
             };
 
             $rootScope.$on('sessionStateChanged', function() {
